Hacky fix to make Reinforce working.
If the arms are initialized with 0 running mean, and same preference, and if the selected arm yields 0 reward,
There will be no update in the running mean and the preference. This would lead to the same arm being picked
again and again.

A hacky fix to solve this issue, was to initialize the running mean with a non-zero number. Though it works now,
Still not sure how correct is this method.